refactor: rename utils dir and shift api dir in
