+++ client/components/layout/index.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react";
 import { Avatar, Button, Dropdown, Layout, Menu } from "antd";
 import styles from "./Layout.module.css";
 import Link from "next/link";
-import { Route, NavMenuKey } from "util/index";
+import { Route, NavMenuKey } from "utils/index";
 import Head from "next/head";
 import { useAuth } from "../authentication";
 import { UserOutlined, CloseCircleFilled } from "@ant-design/icons";
