+++ client/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./api";
 export * from "./constants";
 export * from "./types";
 export * from "./helpers";
